requirements.txt
scapy
scikit-learn
joblib
flask
requests
python-whois
tensorflow